fixed static issue

+++ app/Models/Order.php
@@ -236,7 +236,7 @@ class Order extends Model
 
     public function getGrandTotalCentsAttribute(): int
     {
-        return ($this->grand_total_to_be_paid) * 100;
+        return (int) (($this->grand_total_to_be_paid) * 100);
     }
 
     public function getGrandTotalToBePaidAttribute(): float
